+++ app/Http/Controllers/ReceitaController.php
@@ -25,24 +25,33 @@ class ReceitaController extends Controller
         if ($this->receitaJaExiste($request)) {
             return Responser::error(409, "Já existe uma Receita com essa descrição para o mês informado");
         }
+        $dados['user_id'] = $request->user()->id;
         $receita = Receita::create($dados);
         return $receita;
     }
 
-    public function index()
+    public function index(Request $request)
     {
-        $paginado = Receita::paginate(env('PER_PAGE', 15));
-        return Paginador::paginar($paginado);
+        $dados = $request->all();
+        if (!in_array('descricao', array_keys($dados))) {
+            $result = Receita::with('user')->where('user_id', $request->user()->id)->paginate(env('PER_PAGE', 15));
+            return Paginador::paginar($result);
+        }
+        $result = Receita::with('user')->where('descricao', 'like', "%{$dados['descricao']}%")->where('user_id', $request->user()->id);
+        return Paginador::paginar($result->paginate(env('PER_PAGE', 15)));
     }
 
-    public function show(int $id)
+    public function show(Request $request, int $id)
     {
-        return Receita::findOrFail($id);
+        return Receita::where('user_id', $request->user()->id)->findOrFail($id);
     }
 
     public function update(int $id, Request $request)
     {
         $existente = Receita::findOrFail($id);
+        if ($existente->user->id != $request->user()->id) {
+            return Responser::error(403, "Receita não pertence ao usuário!");
+        }
         $rules = [
             'descricao' => 'required',
             'valor' => 'required|numeric',
@@ -57,8 +66,11 @@ class ReceitaController extends Controller
         return $existente;
     }
 
-    public function destroy(Receita $receita)
+    public function destroy(Request $request, Receita $receita)
     {
+        if ($receita->user->id != $request->user()->id) {
+            return Responser::error(403, "Despesa não pertence ao usuário!");
+        }
         $copia = clone $receita;
         $receita->delete();
         return response()->json(["message" => "Receita apagada com sucesso!", "entidade" => $copia]);
@@ -67,7 +79,13 @@ class ReceitaController extends Controller
     private function receitaJaExiste(Request $request)
     {
         $date = Carbon::createFromFormat("d/m/Y", $request->data);
-        $noBd = Receita::whereMonth('data', $date->month)->whereYear('data', $date->year)->where('descricao', $request->descricao)->get();
+        $noBd = Receita::whereMonth('data', $date->month)->whereYear('data', $date->year)->where('descricao', $request->descricao)->where('user_id', $request->user()->id)->get();
         return count($noBd) > 0;
+    }
+    public function porMes(Request $request, int $ano, int $mes)
+    {
+        $date = Carbon::createFromFormat("d/m/Y", "01/" . $mes . "/" . $ano);
+        $noBd = Receita::whereMonth('data', $date->month)->whereYear('data', $date->year)->where('user_id', $request->user()->id);
+        return Paginador::paginar($noBd->paginate(env('PER_PAGE', 15)));
     }
 }
